Bootstub: Increase the command line size in bootstub
BZ: 12336

Bootstub is currently allocated only 256 bytes for kernel boot command-line
options. The Android kernel boot command line options is getting larger and
with the current size of 256 bytes in size, it's not enough.

This patch added to increase the allocation of the kernel boot command-line
options to 1024 bytes in bootstub. Also, bump up the bootstub version to v1.2

Change-Id: Id8804359899ee1facb621de2cb8fd3c6ef4e14ad
Signed-off-by: Jeremy Compostella <jeremyx.compostella@intel.com>
Signed-off-by: Leonard Mai <leonard.mai@intel.com>
Reviewed-on: http://android.intel.com:8080/22416
Reviewed-by: Du, Alek <alek.du@intel.com>
Reviewed-by: Romieu, Benoit <benoit.romieu@intel.com>
Tested-by: Seibel, Eric <eric.seibel@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>