bootstub support for aosp and legacy
The code searches for the AOSP header at a specified location.
If found, it performs all the required memcpy by using the
addresses and size provided in the AOSP header instead of those
hardcoded into bootstub.h.
Otherwise, it processes as usual for an OSII image.
Change the bootstub output to $(PRODUCT_OUT)/2ndbootloader

Change-Id: I25da96f99b2a01f457d3f8b8393080f9ec562c94
Signed-off-by: Florent Auger <florent.auger@intel.com>
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Signed-off-by: Jocelyn Falempe <jocelyn.falempe@intel.com>
Reviewed-on: https://android.intel.com/215921
Reviewed-by: Betous, GuillaumeX <guillaumex.betous@intel.com>